refactor: 쉽게 관리할 수 있도록 패키지 구조 리팩토링
