+++ src/test/java/carRacing/step1/SetTest.java
@@ -1,4 +1,4 @@
-package study;
+package carRacing.step1;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
