+++ src/test/java/carRacing/step1/StringTest.java
@@ -1,4 +1,4 @@
-package study;
+package carRacing.step1;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
